Fix handling of VCF file with no records (only header).
Before this commit, if no records are in VCF, then parse_header function
would return one of header lines instead of the first record (because
STDIN.each_line loop exits), this results in error message:
    Unexpected line ##fileformat=VCFv4.0
Since it tries to read this line as a record line.

We made an additional check to figure out the line to return from
parse_header is indeed a record line.